backend/requirements.txt
Flask==1.1.4
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.1
Flask-CORS==3.0.10
Flask-JWT-Extended==3.25.1
PyMySQL==0.10.1
python-dotenv==0.18.0
requests==2.25.1
cryptography==3.3.2
Werkzeug==1.0.1